% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{double_centre}
\alias{double_centre}
\title{Double centring of a squared dissimilarity matrix}
\usage{
double_centre(d)
}
\arguments{
\item{d}{A \code{dist} object or a symmetric numeric matrix.}
}
\value{
The Gower centred matrix.
}
\description{
Double centring of a squared dissimilarity matrix
}
