% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference-procrustes.R
\name{protest}
\alias{protest}
\title{Procrustes comparison of two proximity matrices}
\usage{
protest(px1, px2, k = 2L, n_perm = 999)
}
\arguments{
\item{px1, px2}{\code{proximity} objects on the same \code{n} observations.}

\item{k}{Number of MDS dimensions to retain.}

\item{n_perm}{Number of permutations for the PROTEST test.}
}
\value{
An object of class \code{htest}.
}
\description{
Superimposes the classical multidimensional scaling configurations of two
proximity matrices and reports the residual sum of squares \eqn{m^2},
together with the PROTEST permutation test of its significance.
}
\details{
Not implemented yet: scheduled for phase F2.
}
